+++ src/pinkerton/modules/secret.py
@@ -5,11 +5,17 @@ from rich import print
 from src.pinkerton.settings import props
 
 import re
+import jsbeautifier
 
 regex_list = {
     'Google API': r'AIza[0-9A-Za-z-_]{35}',
     'Firebase Key': r'AAAA[A-Za-z0-9_-]{7}:[A-Za-z0-9_-]{140}',
     "Firebase URL": r".*firebaseio\.com",
+    "PGP Private Key Block": r"-----BEGIN PGP PRIVATE KEY BLOCK-----",
+    "SSH (DSA) Private Key": r"-----BEGIN DSA PRIVATE KEY-----",
+    "SSH (EC) Private Key": r"-----BEGIN EC PRIVATE KEY-----",
+    "SSH (RSA) Private Key": r"-----BEGIN OPENSSH PRIVATE KEY-----",
+    "SSH (ssh-ed25519) Public Key": r"ssh-ed25519",
     'Google Captcha Key': r'6L[0-9A-Za-z-_]{38}|^6[0-9a-zA-Z_-]{39}$',
     "Amazon AWS Access Key ID": r"AKIA[0-9A-Z]{16}",
     "Amazon MWS Auth Token": r"amzn\\.mws\\.[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}",
@@ -53,17 +59,12 @@ regex_list = {
     "Square OAuth Secret": r"sq0csp-[0-9A-Za-z\\-_]{43}",
 }
 
-def return_regex_list() -> None:
-    " Extract each regex pattern from regex_list list "
-
-    for pattern_name, pattern_value in regex_list.items():
-        return(pattern_name, pattern_value)
-
 def direct_scan(link) -> None:
     " Open JavaScript file without parsing URL before requesting "
     
     response: function = get(link, **props)
     content: str = response.text
+    content: str = jsbeautifier.beautify(content)
 
     for key, value in regex_list.items():
         pattern = value
@@ -79,6 +80,7 @@ def passed_scan(final_url):
 
     response: function = get(final_url, **props)
     content: str = response.text
+    content: str = jsbeautifier.beautify(content)
 
     for key, value in regex_list.items():
         pattern = value